requirements.txt
aniso8601==1.1.0
argon2-cffi==16.2.0
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
cffi==1.8.3
click==6.6
coverage==4.2
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-RESTful==0.3.5
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
more-itertools==4.2.0
peewee==2.10.2
pluggy==0.6.0
py==1.5.4
pycparser==2.18
pytest==3.6.3
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
Werkzeug==0.14.1
WTForms==2.2.1
